Add options WITH_SSL and WITH_ZLIB
Mysql-connector-c includes an embedded SSL library (yassl) and an embedded
Zlib. WITH_SSL and WITH_ZLIB can be set to "system" to link with system
versions of OpenSSL and Zlib. This patch allows these options to be set
and then adds dependencies on the Conan.io OpenSSL and Zlib packages.